requirements.txt
numpy==1.26.4
pillow==10.2.0
matplotlib==3.8.2
pycocotools==2.0.7
tensorboard==2.14.0
torch==2.9.1
torchvision==0.24.1
transformers==4.42.4
scipy==1.11.4
albumentations==1.3.1
opencv-python-headless==4.9.0.80
tqdm==4.66.1
